+++ pa11yci.config.js
@@ -14,7 +14,7 @@ const CfAccessClientId = process.env.CF_ACCESS_CLIENT_ID;
 const CfAccessClientSecret = process.env.CF_ACCESS_CLIENT_SECRET;
 if (!isLocalHost && (!CfAccessClientId || !CfAccessClientSecret)) {
   throw new TypeError(
-    "Please specify Cloudflare Access token headers in envs\nfor background info see https://developers.cloudflare.com/cloudflare-one/identity/service-tokens/",
+    "Please specify Cloudflare Access token headers in envs\nfor background info see https://developers.cloudflare.com/cloudflare-one/identity/service-tokens/"
   );
 }