+++ src/components/SearchFilters/ActiveFilters.test.tsx
@@ -55,10 +55,10 @@ describe("ActiveFilters", () => {
   test.each(
     searchFilters.keyStageFilters
       .filter((ks) => ks.checked)
-      .map((ks) => ks.title),
+      .map((ks) => ks.title)
   )("should render the checked filters: %s", (ks) => {
     const { getByRole } = renderWithTheme(
-      <ActiveFilters searchFilters={searchFilters} />,
+      <ActiveFilters searchFilters={searchFilters} />
     );
     const button = getByRole("button", { name: `Remove ${ks} filter` });
     expect(button).toBeInTheDocument();
@@ -66,10 +66,10 @@ describe("ActiveFilters", () => {
   test.each(
     searchFilters.keyStageFilters
       .filter((ks) => !ks.checked)
-      .map((ks) => ks.title),
+      .map((ks) => ks.title)
   )("should not render the unchecked filters: %s", (ks) => {
     const { queryByRole } = renderWithTheme(
-      <ActiveFilters searchFilters={searchFilters} />,
+      <ActiveFilters searchFilters={searchFilters} />
     );
     const button = queryByRole("button", { name: `Remove ${ks} filter` });
     expect(button).not.toBeInTheDocument();
@@ -77,10 +77,10 @@ describe("ActiveFilters", () => {
   test.each(
     searchFilters.subjectFilters
       .filter((subject) => subject.checked)
-      .map((subject) => subject.title),
+      .map((subject) => subject.title)
   )("should render the checked filters: %s", (subject) => {
     const { getByRole } = renderWithTheme(
-      <ActiveFilters searchFilters={searchFilters} />,
+      <ActiveFilters searchFilters={searchFilters} />
     );
     const button = getByRole("button", { name: `Remove ${subject} filter` });
     expect(button).toBeInTheDocument();
@@ -89,10 +89,10 @@ describe("ActiveFilters", () => {
   test.each(
     searchFilters.subjectFilters
       .filter((subject) => !subject.checked)
-      .map((subject) => subject.title),
+      .map((subject) => subject.title)
   )("should not render the unchecked filters: %s", (subject) => {
     const { queryByRole } = renderWithTheme(
-      <ActiveFilters searchFilters={searchFilters} />,
+      <ActiveFilters searchFilters={searchFilters} />
     );
     const button = queryByRole("button", { name: `Remove ${subject} filter` });
     expect(button).not.toBeInTheDocument();
@@ -101,10 +101,10 @@ describe("ActiveFilters", () => {
   test.each(
     searchFilters.contentTypeFilters
       .filter((type) => type.checked)
-      .map((ContentType) => ContentType.title),
+      .map((ContentType) => ContentType.title)
   )("should render the checked type filters: %s", (ContentType) => {
     const { getByRole } = renderWithTheme(
-      <ActiveFilters searchFilters={searchFilters} />,
+      <ActiveFilters searchFilters={searchFilters} />
     );
     const button = getByRole("button", {
       name: `Remove ${ContentType} filter`,
@@ -115,10 +115,10 @@ describe("ActiveFilters", () => {
   test.each(
     searchFilters.contentTypeFilters
       .filter((type) => !type.checked)
-      .map((ContentType) => ContentType.title),
+      .map((ContentType) => ContentType.title)
   )("should not render the unchecked type filters: %s", (ContentType) => {
     const { queryByRole } = renderWithTheme(
-      <ActiveFilters searchFilters={searchFilters} />,
+      <ActiveFilters searchFilters={searchFilters} />
     );
     const button = queryByRole("button", {
       name: `Remove ${ContentType} filter`,